activities with validation, missing sub and unsub from activity

+++ controllers/activities.controller.js
@@ -6,21 +6,25 @@ exports.create = async (req, res) => {
   const date = new Date();
   const todayDate = date.getFullYear() + "-" + ('0' + (date.getMonth() + 1)).slice(-2) + "-" + ('0' + date.getDate()).slice(-2);
   try {
-      //missing validate user that creates
-      console.log(req.loggedUser)
+    if(req.loggedUser.role == 'admin') {
       if(req.body.start <= todayDate || req.body.start >= req.body.end || req.body.end <= todayDate)
         return res.status(400).json({
           success: false,
           msg: 'Please provide a valid date'
         })
       let activity = await Activity.create(req.body)
-      res.status(201).json({
+      return res.status(201).json({
         success: true,
-        message: 'Activity successful created!',
-        URL: `/activities/${activity.id}`,
+        message: `Activity successful created!`,
+        URL: `/activities/${activity.id}`
       });
+    }
+    res.status(400).json({
+      succes: false,
+      message: 'Invalid credentials'
+    })
     } catch (err) {
-      if (err instanceof ValidationError) // Tutorial model as validations for title and published
+      if (err instanceof ValidationError)
         res.status(400).json({ success:false, msg: err.errors.map(e => e.message) });
       else
         res.status(500).json(
@@ -68,51 +72,70 @@ exports.findOne = async (req, res) => {
   }
 }
 exports.edit = async (req, res) => {
-  //verify if user is valid, like in create
   try {
-    let activity = await Activity.findByPk(req.params.idA)
-    if(activity === null){
-      return res.status(404).json({
-        sucess: false,
-        msg: `Activity not found.`
-      })
-    }
-    Activity.update(
-      {name: req.body.name},
-      {description: req.body.description},
-      {start: req.body.start},
-      {end: req.body.end},
-      {location: req.body.location},
-      {image: req.body.image},
-      {
-        where: {id: req.params.idA}
+    if(req.loggedUser.role == 'admin'){
+      let activity = await Activity.findByPk(req.params.idA)
+      if(activity === null){
+        return res.status(404).json({
+          sucess: false,
+          msg: `Activity not found.`
+        })
       }
-      )
-      res.status(202).json({
+      if (req.body.name) activity.name = req.body.name;
+      if (req.body.description) activity.description = req.body.description;
+      if (req.body.start) activity.start = req.body.start;
+      if (req.body.end) activity.end = req.body.end;
+      if (req.body.location) activity.location = req.body.location;
+      if (req.body.image) activity.image = req.body.image;
+      Activity.update(
+        {name: req.body.name,
+        description: req.body.description,
+        start: req.body.start,
+        end: req.body.end,
+        location: req.body.location,
+        image: req.body.image},
+        {
+          where: {id: req.params.idA}
+        }
+        )
+      return res.status(202).json({
         succes: true,
         msg: `Activity with ID ${activity.id} updated successfully`
       })
-        
-        
-      } catch (err) {
+    }
+    res.status(400).json({
+      succes: false,
+      message: 'Invalid credentials'
+    })
+  } catch (err) {
     res.status(500).json({
       success: false,
       msg: err.message || 'Some error occurred while creating a new user.'
     })
   }
 }
-exports.delete = async (req, res, next) => {
+exports.delete = async (req, res) => {
   try {
-    let result = await Activity.destroy({
-        where: {id: req.params.idA}
-    })
-    if(result == 1) {
-        res.status(200).json({
-            success: true,
-            msg: `Deleted Activity with ID ${req.params.idA} successfully`
+    if(req.loggedUser.role == 'admin'){
+      let result = await Activity.destroy({
+          where: {id: req.params.idA}
+      })
+      if(result == 0) {
+        return res.status(404).json({
+            success: false,
+            msg: `Activity with ID ${req.params.idA} Not Found`
         });
+      }
+      res.status(200).json({
+          success: true,
+          msg: `Deleted Activity with ID ${req.params.idA} successfully`
+      });
     }
-} catch (err) {
+    res.status(400).json({
+        success: false,
+        msg: 'Invalid credentials'
+    });
+  } catch (err) {
     res.status(500).json({
         success: false,
         msg: err.message || `Some error occurred while getting the activity with ID ${req.params.idA}.`
